refactor: extract context to different file
