+++ frontend/src/features/admin-form/create/common/CreatePageSidebarContext/CreatePageSidebarContext.tsx
@@ -7,10 +7,8 @@ import {
   useMemo,
   useState,
 } from 'react'
-import { UseMeasureRef } from 'react-use/lib/useMeasure'
 
 import { useIsMobile } from '~hooks/useIsMobile'
-import { useMeasure } from '~hooks/useMeasure'
 
 import { FieldListTabIndex } from '../../builder-and-design/constants'
 import {
@@ -45,19 +43,9 @@ type CreatePageSidebarContextProps = {
   isDrawerOpen: boolean
   fieldListTabIndex: FieldListTabIndex
   setFieldListTabIndex: (tabIndex: FieldListTabIndex) => void
-  // drawerRef: UseMeasureRef<HTMLDivElement>
-  // drawerWidth: number
-}
-type CreatePageSidebarWidthContextProps = {
-  drawerRef: UseMeasureRef<HTMLDivElement>
-  drawerWidth: number
 }
 const CreatePageSidebarContext = createContext<
   CreatePageSidebarContextProps | undefined
->(undefined)
-
-const CreatePageSidebarWidthContext = createContext<
-  CreatePageSidebarWidthContextProps | undefined
 >(undefined)
 
 export const useCreatePageSidebar = (): CreatePageSidebarContextProps => {
@@ -69,17 +57,6 @@ export const useCreatePageSidebar = (): CreatePageSidebarContextProps => {
   }
   return context
 }
-export const useCreatePageSidebarWidth =
-  (): CreatePageSidebarWidthContextProps => {
-    const context = useContext(CreatePageSidebarWidthContext)
-    if (!context) {
-      throw new Error(
-        `useCreatePageSidebar must be used within a CreatePageSidebarWidthProvider component`,
-      )
-    }
-    return context
-  }
-
 export const useCreatePageSidebarContext =
   (): CreatePageSidebarContextProps => {
     const isMobile = useIsMobile()
@@ -187,17 +164,6 @@ export const useCreatePageSidebarContext =
       handleClose,
       fieldListTabIndex,
       setFieldListTabIndex,
-      // drawerRef,
-      //     drawerWidth,
-    }
-  }
-
-export const useCreatePageSidebarWidthContext =
-  (): CreatePageSidebarWidthContextProps => {
-    const [drawerRef, { width: drawerWidth }] = useMeasure<HTMLDivElement>()
-    return {
-      drawerRef,
-      drawerWidth,
     }
   }
 
@@ -212,14 +178,5 @@ export const CreatePageSidebarProvider: FC = ({ children }) => {
     <CreatePageSidebarContext.Provider value={context}>
       {children}
     </CreatePageSidebarContext.Provider>
-  )
-}
-
-export const CreatePageSiderBarWidthProvider: FC = ({ children }) => {
-  const context = useCreatePageSidebarWidthContext()
-  return (
-    <CreatePageSidebarWidthContext.Provider value={context}>
-      {children}
-    </CreatePageSidebarWidthContext.Provider>
   )
 }
